Adds image whitelisting and blacklisting feature in policy bundle.
Adds 'whitelisted_images' and 'blacklisted_images' properties to the policy bundle that are each arrays of mapping rules
that specify image match parameters. An image match in the whitelist will cause the evaluation to be 'go'.  A blacklist
match will cause the result to be 'stop'. Blacklist matches override whitelist matches.

Also adds a 'final_action_reason' in the bundle evaluation output that can have value 'blacklisted', 'whitelisted', or 'policy_evaluation'.

Signed-off-by: Zach Hill <zach@anchore.com>