Update anchore-engine startup script to only run services found in config as enabled.
Instead of trying every service and having the service exit if it finds itself not enabled in configuration,
the startup script reads the configuration and only attempts to start enabled services. This is overridden if the
ANCHORE_ENGINE_SERVICES environment variable is set and will use that value if present.

Signed-off-by: Zach Hill <zach@anchore.com>