Adds support for multiple policies in mapping rules of policy bundles.
Adds 'policy_ids' property to mapping rule as alternative to 'policy_id' property,
such that either is supported and if both are found or neither are found, and error is raised.

Bundle execution is done in policy_id order with no short-circuit or deduplication of results.

Adds a new final column to the policy evaluation output in table-form: 'policy_id' to map each result row to the policy
that generated it.

Signed-off-by: Zach Hill <zach@anchore.com>