Revert "policy-engine - use full version and not the epochless version of fixed package for version comparison"
This reverts commit dfa717d2399206fe04d705e6777a3a16d46777b5.

Original commit changed the attribute of the fixed package used for version comparison from epoch-less to fully qualified version. This change was prompted by a CVE in ubuntu namespace that was not matched correctly - its a case where installed package has a non-0 epoch but the epoch on the fixed version is ignored and assumed to be 0. While the original commit fixes this issue in ubuntu namespace, it has the opposite effect in RHEL based distros. Epoch is not gathered/stored for RHEL packages. To complicate the issue, a significant portion of fixed versions in RHEL have non-0 epochs. In combination with the original commit this would result in not matching CVEs correctly, hence the revert. The following changes are required in order to fix all of this correctly
1. Change analyzers to gather epoch information if available
2. Undo this revert and put back the original commit
3. Additionally update the version comparators to drop non-0 epoch on fixed versions if the installed version has no epoch information