Adds support for specifying registry credentials for specific repositories or sets of repos using wildcards.
To use, in the POST /registries request, the 'registry' property value can now include a '/' as well as a trailing wildcard (*).

Examples: "docker.io/library*", "gcr.io/myproject*", "gcr.io/myproject/*", "docker.io/myaccount/*"

Anchore treats anything with a '/' as a registry, repository pair and performs either an exact or prefix match depending on if the trailing '*' is found. When the system needs credentials to pull an image it will search for matches using the longest registry entry first and checking each for a match in descending length order (to ensure a longest-prefix match). If one matches, it is used.

Implements #142

Signed-off-by: Zach Hill <zach@anchore.com>