Rename anchore-manager archivestorage command to objectstorage, and add cli options for src/dest using analysis_archive
Supports specifying a source and/or dest configured by the analysis_archive section. If the config is not found, the defaults are not used, but will error out indicating that you should use the non-archive-section migration path.

NOTE: this migrates all object storage data using that backend, not just analysis archive data if they are mixed. The cli options simply indicate with backend configurations to migrate between, but do not filter the objects themselves which are migrated to only those in analysis archive, this is important if migrating a single-backend to/from a multi-backend system.

Future updates will support migration of specific objects/buckets (e.g. the analysis_archive objects)

Signed-off-by: Zach Hill <zach@anchore.com>