Adds support for oauth tokens issued by the api service from POST /v1/oauth/token. Also adds storage of passwords via secure hash in db.
Token issuance uses the Resource owner password grant flow with an anonymous client option. Tokens are jwt tokens that
are signed using either the private key of pub/private pair or a shared key.

Configuration of the keys is in the config.yaml. Adds a new 'keys' config section like:
    keys:
      secret: <somesecre>
      private_key_path: <path to private key as pem>
      public_key_path: <path to public key as pem>

User authentication is configured also in config.yaml with a 'user_authentication' section:

    user_authentication:
      oauth:
        enabled: <bool, default false>
        token_expiration_seconds: <int, default 3600>
      hashed_passwords: <bool, default=false>

The 'hashed_passwords' option will enable strong hashes (default's to argon2) for passwords stored in db.

Signed-off-by: Zach Hill <zach@anchore.com>