tests: remove fork() to run flask app
Swagger Validation can be caught by running the flask app as a test
client. This is the same way the ``connexion`` library tests internally.
By forking it would mess up pytest runs, and failures on setup wouldn't
be noticed.

This change introduces fixtures so that each item in the swagger path is
represented as a test, increasing visibility into what passes or fails.

Signed-off-by: Alfredo Deza <adeza@anchore.com>