Modify behavior of the lookup process for artifact namespaces to be more strict.
Old behavior: use distro mappings's to_distro field to get set of distros mapped to a candidate, but filter that list to
remove any distros that have an exact match in the feeds. This was originally due to distro-mappings being code rather than data,
so gave flexibility to add new feed groups and have it automatically picked up.

New behavior: same as previous with sole exception that no filtering is done. The distro-mapping records define the exact mapping
regardless of feed groups. This is acceptable now because there is an API for that mapping data so new groups can be supported
by explicit API calls to add new mappings.

This change was required to ensure clean transition of centos:* namespace to rhel:* namespace for matches and ensuring that
centos-distro images don't fall back on their old namespace despite the name match. The earlier solution to this was to use
the enabled status field on the group, but that should control sync behavior only, not matching behavior to remain clear.

Signed-off-by: Zach Hill <zach@anchore.com>