Reverts change to using package version exact match as fallthru in FixedArtifact matches.
This is correct but the case it addresses: non-ranged vuln and fix records needs to be handled
with a larger change to how the two types of records intereact. For now this resolves incorrect
results due to it always matching even when VulnerableArtifact records are present and should be used instead.

Signed-off-by: Zach Hill <zach@anchore.com>