Do not build Skopeo binary twice
bin/skopeo is a PHONY target, which is also a dependency for
install-binary target in
https://github.com/containers/skopeo/tree/master/Makefile. This means,
that results of bin/skopeo are not tracked by make and build was called
twice: first when calling `make bin/skopeo` and then as a dependency of
`make install-binary`.

This commit eliminates duplicate building of Skopeo binary.

Signed-off-by: Oleksandr Kovalchuk <anxolerd@outlook.com>