Use OCI manifest from skopeo copy for analysis
If layers are not compressed on the registry, skopeo copy will compress them
on the fly when executing "skopeo copy ... oci:dir".  That causes disparity
between image manifest that was previously fetched with "skopeo inspect",
which had references to uncompressed layers.

This change makes analysis rely on the OCI image manifest only.

Fixes #1080

Signed-off-by: Tero Saarni <tero.saarni@est.tech>