Internal (catalog) API for returning all or multiple types of content for an image.
There is a similar API in place but it returns content for a single type per request. This is a convenience API to avoid multiple round trips and repetitive processing in catalog. It returns content for multiple types or all types depending on the query params. The implementation is an extension of the existing API.

Signed-off-by: Swathi Gangisetty <swathi@anchore.com>