Fixes image import finalization state to be 'processing' instead of 'complete'.
Using 'complete' could cause an import to be GC'd before the analyzer
actually finishes the task. The analyzer also updates status to
'complete', so this was as bug where it should have been 'processing'
state instead.

Expiration logic still is active on the 'processing' state but with a
longer ttl as expected.

Signed-off-by: Zach Hill <zach@anchore.com>