Adds the master-branch side of the migration of docs content to the docs branch only.
This removes the docs/ directory, submodule for docsy, and the gihtub
action workflow for publication.

All of that content and CI functions are in the "docs" and
"publish-docs" branches now.

Future updates to the docs should branch from and merge into the "docs"
branch. Publication should be just merging docs branch into publish-docs
branch.

Signed-off-by: Zach Hill <zach@anchore.com>